Release v0.31
Based on suggestions below, make the documentation
cleaner.

Fixed - The script URL is broken in your CONTRIBUTING docs
Untitled

Commented - And this statement, If git is already installed you need to
put "C:\Program Files\Git\bin" first on your PATH. - do you mean the
Git\bin needs to be at the top of the list, or just listed in the
environmental variable?